Chashi Bhai/requirements.txt
# Core Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
starlette>=0.41.0

# HTTP Client
httpx>=0.26.0

# Environment & Configuration
python-dotenv>=1.0.0

# AI & LangChain
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-groq>=0.1.0
langchain-community>=0.0.10
openai>=1.7.0

# Vector Database
chromadb>=0.4.24

# Search Tools
arxiv>=2.3.0
wikipedia>=1.4.0
duckduckgo-search>=4.0.0

# Translation & Language Detection
deep-translator>=1.11.4
langdetect>=1.0.9

# Data Validation
pydantic>=2.5.0
